[NO-TICKET] Organize doc-site components into folders (#2159)
* Organize doc site components into folders now that there are a lot of them

* Fix broken link on index page

* Rename folder to be more generic

* Fix type export

* Rename folders to lowercase
