+++ packages/docs/content/not-in-sidebar/intro.mdx
@@ -2,7 +2,7 @@
 title: Introduction
 ---
 
-The design system is a set of open source design and front-end development resources for creating Section 508 compliant, responsive, and consistent websites. It builds on the U.S. Web Design System and extends it to support additional CSS and React components, [utility classes](/utilities/overview/), and a [grid framework](/utilities/layout-grid/) to allow teams to quickly prototype and build accessible, responsive, production-ready websites.
+The design system is a set of open source design and front-end development resources for creating Section 508 compliant, responsive, and consistent websites. It builds on the U.S. Web Design System and extends it to support additional CSS and React components, [utility classes](/utilities/overview/), and a [grid framework](/foundation/layout-grid/layout-grid/) to allow teams to quickly prototype and build accessible, responsive, production-ready websites.
 
 It is currently being applied to [HealthCare.gov](https://www.healthcare.gov/) and [Medicare.gov](https://www.medicare.gov/). It is open-source and freely available to use by anyone.
