+++ packages/docs/src/components/content/EmbeddedExample.tsx
