+++ packages/docs/src/pages/404.tsx
@@ -1,10 +1,10 @@
 import * as React from 'react';
 import { graphql } from 'gatsby';
 
-import Layout from '../components/Layout';
+import Layout from '../components/layout/Layout';
 import { MdxQuery } from '../helpers/graphQLTypes';
 import useTheme from '../helpers/useTheme';
-import ContentRenderer from '../components/ContentRenderer';
+import ContentRenderer from '../components/content/ContentRenderer';
 
 const NotFoundPage = ({ data, location }: MdxQuery) => {
   const theme = useTheme();
