+++ packages/docs/src/pages/contact.tsx
@@ -1,10 +1,10 @@
 import * as React from 'react';
 import { graphql } from 'gatsby';
 
-import Layout from '../components/Layout';
+import Layout from '../components/layout/Layout';
 import { MdxQuery } from '../helpers/graphQLTypes';
 import useTheme from '../helpers/useTheme';
-import ContentRenderer from '../components/ContentRenderer';
+import ContentRenderer from '../components/content/ContentRenderer';
 
 const ContactPage = ({ data, location }: MdxQuery) => {
   const { slug } = data.mdx;
